(yeah, no recursion.)
c-with-gc.h:42:1: warning: "ENV_INIT" redefined
c-with-gc.h:41:1: warning: this is the location of the previous definition

Problem is that cpp doesn't expand itself it seems from the macro
output, and yeah can't define it a second time that would be visible
beforehand. fah. (And can't include a macro definition in the output
either, even if that helped.)